// src/users-levels/enums/user-level.enum.ts
export enum UserLevelEnum {
	Beginner = 'Beginner', // 0 - 100 points
	Novice = 'Novice', // 101 - 300 points
	Learner = 'Learner', // 301 - 600 points
	Skilled = 'Skilled', // 601 - 1000 points
	Advanced = 'Advanced', // 1001 - 1500 points
	Expert = 'Expert', // 1501 - 2100 points
	Master = 'Master', // 2101 - 2800 points
	Grandmaster = 'Grandmaster', // 2801 - 3600 points
	Legend = 'Legend', // 3601 - 4500 points
	Immortal = 'Immortal' // 4501+ points
}